Move statics from `ExportPlugin` to utils class
