+++ src/CMakeLists.txt
@@ -592,6 +592,8 @@ list( APPEND SOURCES
 
       export/Export.cpp
       export/Export.h
+      export/ExportUtils.cpp
+      export/ExportUtils.h
 
       # Standard exporters
       export/ExportCL.cpp
