+++ src/export/Export.cpp
@@ -119,58 +119,7 @@ void ExportPlugin::OptionsCreate(ShuttleGui &S, int WXUNUSED(format))
    S.EndHorizontalLay();
 }
 
-//Create a mixer by computing the time warp factor
-std::unique_ptr<Mixer> ExportPlugin::CreateMixer(const TrackList &tracks,
-         bool selectionOnly,
-         double startTime, double stopTime,
-         unsigned numOutChannels, size_t outBufferSize, bool outInterleaved,
-         double outRate, sampleFormat outFormat,
-         MixerSpec *mixerSpec)
-{
-   Mixer::Inputs inputs;
 
-   bool anySolo = !(( tracks.Any<const WaveTrack>() + &WaveTrack::GetSolo ).empty());
-
-   auto range = tracks.Any< const WaveTrack >()
-      + (selectionOnly ? &Track::IsSelected : &Track::Any )
-      - ( anySolo ? &WaveTrack::GetNotSolo : &WaveTrack::GetMute);
-   for (auto pTrack: range)
-      inputs.emplace_back(
-         pTrack->SharedPointer<const SampleTrack>(), GetEffectStages(*pTrack));
-   // MB: the stop time should not be warped, this was a bug.
-   return std::make_unique<Mixer>(move(inputs),
-                  // Throw, to stop exporting, if read fails:
-                  true,
-                  Mixer::WarpOptions{tracks},
-                  startTime, stopTime,
-                  numOutChannels, outBufferSize, outInterleaved,
-                  outRate, outFormat,
-                  true, mixerSpec);
-}
-
-void ExportPlugin::InitProgress(std::unique_ptr<BasicUI::ProgressDialog> &pDialog,
-   const TranslatableString &title, const TranslatableString &message)
-{
-   if (!pDialog)
-      pDialog = std::make_unique<ProgressDialog>( title, message );
-   else
-   {
-      if (auto pd = dynamic_cast<ProgressDialog*>(pDialog.get()))
-      {
-         pd->SetTitle(title);
-         pd->Reinit();
-      }
-
-      pDialog->SetMessage(message);
-   }
-}
-
-void ExportPlugin::InitProgress(std::unique_ptr<BasicUI::ProgressDialog> &pDialog,
-   const wxFileNameWrapper &title, const TranslatableString &message)
-{
-   return InitProgress(
-      pDialog, Verbatim( title.GetName() ), message );
-}
 
 //----------------------------------------------------------------------------
 // Export
